backend/requirements.txt
Django==3.1
djangorestframework
django-filter
celery
redis
drf-yasg==1.20.0
psycopg2-binary